Fix Windows build errors
